# Optimize IntMap.Bin