--- a/containers-tests/containers-tests.cabal
+++ b/containers-tests/containers-tests.cabal
@@ -58,7 +58,6 @@ common benchmark-deps
     , tasty-bench       >=0.3.1   && <0.4
 
   -- Flags recommended by tasty-bench
-  ghc-options:      "-with-rtsopts=-A32m"
   if impl(ghc >= 8.6)
     ghc-options:    -fproc-alignment=64