… containers-tests/benchmarks/LookupGE/LookupGE_IntMap.hs
… containers-tests/tests/IntMapValidity.hs
… containers/src/Data/IntMap/Strict/Internal.hs