@meooow25 2024-03-10
@treeowl I assume you mean #994. I don't know much about the CI setup but I'll take a look.

> With a key and a compatible prefix, xor only yields a value at or below the level of the masking bit, whether or not the value is at the masking bit determines the direction (0 is right).

@BurningWitness it is unclear to me how that would work. For instance, `xor p k .&. p == 0`, which is the same as `k .&. p == p`, only tells us whether `p`'s 1-bits agree with `k`. But we also need to know if it the 0-bits agree.  
Probably worth thinking about it though, it would be great to find a more efficient set of bitwise ops to do the job.