@BurningWitness 2024-03-10
Oh, true, `xor` only has that guarantee if prefix matches the key, so I was wrong.