@meooow25 2024-03-10
> Oh, true, `xor` only has that guarantee if prefix matches the key, so I was wrong.

Ah well if the prefix already matches, the branch into left or right is done with one comparison so it's hard to imagine something faster than that.  
What I'd like to find a better way for is the `nomatch` stuff.