@meooow25 2024-03-11
I believe `nomatch` should be pretty much as fast as the old one now!

Previous: ``i .&. (m `xor` (-m)) == p`` — 1 and, 1 xor, 1 negate, 1 compare
New: ``(i `xor` p) .&. (p `xor` (-p))`` — 1 and, 2 xor, 1 negate

Correct me if I'm wrong, I know little about things at this low a level.

---

Now, if only `shorter` could be faster or avoided in some manner...
I suspect that's the only thing causing minor slowdowns for the `intersection_disj` case.