@treeowl 2024-03-11
I doubt the comparison costs significantly over the test, but benchmarking is the way to know for sure.