@meooow25 2024-03-12
Okay I found a better way to branch that avoids `shorter`. Almost every benchmark is faster or the same now! 🎉 

@treeowl would you like to review? I think it's in a good shape now.