@BurningWitness 2024-03-12
It would be nice to wrap `p .|. (p - 1)` and `p .&. (p - 1)` into functions like `upper` and `lower` respectively for clarity.

---

Also this should mean precise lookups can be expressed as

```haskell
trimatch p k =
  if k < p
    then if k >= lower p
           then goLeft
           else goFail
           
    else if k <= upper p
           then goRight
           else goFail
```

(I think this is faster than the algorithm listed in the paper, four operations in any direction instead of four for failure plus one for picking a side)