@meooow25 2024-03-13
@BurningWitness 

> It would be nice to wrap `p .|. (p - 1)` and `p .&. (p - 1)` into functions named `upper` and `lower` respectively for clarity.

If that becomes the predominant way of viewing `Prefix`, then perhaps. Currently we also use `nomatch` so that is not true.

> Also this should mean precise lookups can be expressed as...

I like that, since it uses the same view as `mapMapBranch`, but it makes no difference on benchmarks for me. A small concern (maybe?) is that this duplicates the failure branch.  
So we could switch to it, but I don't really see a good reason to do so. If it is demonstrated to be better in some way, then sure I would favor it.