@BurningWitness 2024-03-14
> A small concern (maybe?) is that this duplicates the failure branch.

Indeed that is the case, the failure branch needs to be let-bound with a `NOINLINE` to avoid it. Probably won't be visible on the benchmarks anyway though.