@BurningWitness 2024-03-14
I made a local benchmark on a `Word`-based tree clone and indeed the `nomatch` solution is ahead performance-wise by a remarkably narrow margin, 1-2% difference.