@meooow25 2024-03-29
Awesome 🎉 
I peeked at the Core for `union` and it seemed alright. But I will run the benchmarks a final time. I also want to go over the PR once to check for bugs, since it touches so many functions. I will update when I'm done, should be some time later today.