@meooow25 2024-03-30
Ok I think this is good to go.

Updated the benchmarks. I found some odd increases in a few `intersection` tests that seemed to have originated with 82f4852, a completely unrelated change. Adding a `-fproc-alignment=64` flag (as suggested by `tasty-bench`) fixed that.

I also added the `"-with-rtsopts=-A32m"` flag from tasty-bench while at it. This greatly reduced the `Copied` stats however, for both before and after. I think that column is less indicative of anything now, but I've kept it for the sake of it.

Let me know if thing look good and I'll squash it into a few commits (benchmark flag, tests, main IntMap changes).