@treeowl 2024-03-30
Why would you set `-A32m` here? That seems to change what the benchmarks are measuring quite a lot. The alignment thing seems reasonable, although it's unfortunate we have to do that.