@meooow25 2024-03-30
`tasty-bench` claims it makes the individual benchmarks less likely to affect one another. I thought that would be good to have as a general improvement (and not to address by any observed issue).