@treeowl 2024-03-30
Let's wait with that one for now. I'd like to hear opinions from others on a separate PR.