@meooow25 2024-03-30
Sounds good. Removed the flag and re-ran the benchmarks.